requirements.txt
 Django
 mysql
 django_twilio
 twilio_python